feat: Provide Docker option for setting up backend (#110)
* feat: Provide Docker option for setting up backend

- pass optional --docker flag to wordpress:setup that makes some minor
adjustments to how the setup runs to accommodate Docker. Use localhost
instead of 0.0.0.0 for the db ip since mysql can't connect to 0.0.0.0
within a Docker container. MySQL is installed as part of the Dockerfile
so MySQL doesn't need to be installed during this part of the setup.
- the wordpress server needs to bind to 0.0.0.0 in order for Docker port
binding to work.
- wp needs to be passed --allow-root since in our Docker container we're
running everything as root.

* fixup! feat: Provide Docker option for setting up backend

* style: clean up lint